add http methods to routes

+++ app.py
@@ -1,9 +1,10 @@
 import os
 
-import listener
 from flask import Flask, jsonify, render_template, request
 from flask_cors import CORS
 from flask_socketio import SocketIO
+
+import listener
 from models import Book, db
 
 app = Flask(__name__)
@@ -16,7 +17,7 @@ CORS(app)
 app.config['CORS_HEADERS'] = 'Content-Type'
 db.init_app(app)
 
-thread1 = listener.myThread(1, 1)
+thread1 = listener.myThread("db listener", 1)
 thread1.daemon = True
 thread1.start()
 
@@ -26,7 +27,7 @@ def hello():
     return render_template('index.html')
 
 
-@app.route("/add")
+@app.route("/add", methods=['POST'])
 def add_user():
     name = request.args.get('name')
 
@@ -41,17 +42,20 @@ def add_user():
         return(str(e))
 
 
-@app.route("/del/<id_>")
+@app.route("/del/<int:id_>", methods=['DELETE'])
 def del_user(id_):
-    try:
-        Book.query.filter_by(id=id_).delete()
-        db.session.commit()
-        return "User with id {} deleted".format(id_)
-    except Exception as e:
-        return(str(e))
+    if Book.query.filter_by(id=id_).first():
+        try:
+            Book.query.filter_by(id=id_).delete()
+            db.session.commit()
+            return "User with id {} deleted".format(id_)
+        except Exception as e:
+            return(str(e))
+    else:
+        return "User with id {} does not exist".format(id_)
 
 
-@app.route("/getall")
+@app.route("/getall", methods=['GET'])
 def get_all():
     try:
         books = Book.query.all()
@@ -60,7 +64,7 @@ def get_all():
         return(str(e))
 
 
-@app.route("/get/<id_>")
+@app.route("/get/<int:id_>", methods=['GET'])
 def get_by_id(id_):
     try:
         book = Book.query.filter_by(id=id_).first()
@@ -80,11 +84,8 @@ def handle_connect():
 
 def send_mymsg(msg):
     with app.test_request_context('/'):
-        # socketio.emit('somevent', "{'sendddd_EMIT': 4442}", namespace='/')
         socketio.emit('somevent', msg, namespace='/')
-        socketio.send({'sendddd_SEND': 4442}, namespace='/')
 
 
 if __name__ == '__main__':
-    # socketio.run(app, threaded=True, debug=True)
     socketio.run(app)
